requirements/requirements.txt
django==2.0.1 
djangorestframework==3.7.7
pytz==2017.2
psycopg2==2.7.1
requests==2.18.2
google-auth == 1.3.0
google-api-python-client == 1.6.5
google-auth-oauthlib == 0.2.0
google-auth-httplib2 == 0.0.3